requirements.txt
streamlit
pandas
numpy
scikit-learn>=1.0.0
joblib
matplotlib
PyPDF2
groq
sentence-transformers
faiss-cpu
langgraph
